fix : Repo Test를 위해 JPA -> StringDataJPA로 변경
Repository 테스트를 위해서는 @DataJPATest 로 해야함.
* 하지만 코드는 JPA로 구현했음.
* Repository 테스트 진행을 위해서 SpringDataJPA를 사용해야함.

SpringDataJPA 사용으로 인한 코드 변경 -> save
* Reposiotry save method를 id만 반환하도록 만들었음.
* Spring Data JPA는 save method를 전체 반환으로 제작함.
* 그에 맞도록 코드를 수정

SpringDataJPA 사용으로 인한 코드 변경 -> count
* count로 JPA가 지원하는 갯수 반환 interface 구현체가 있음. 이를 사용함

SpringDataJPA 사용으로 인한 코드 변경 -> currentDate 외부로 제한
* SrpingDataJPA를 사용하면서 쿼리만 Repo에 들어가야함.
* 해당 로직은 Service에서 처리

SpringDataJPA 사용으로 인한 코드 수정 불가 -> api 중 일부 구현 실패
* 마감기간이 가장 빠른  4개 Mission 제공 구현 실패
* 가장 인기있는 4개 Mission 제공 구현 실패
* 사유는, 이전에는 내부에 연산로직을 포함했는데, 이번엔 그럴 수 없음.
* 추후 구현예정

Resolves: #issueNo
See also: #issueNo

+++ src/main/java/io/wisoft/testermatchingplatform/web/dto/response/CountResponse.java
@@ -7,16 +7,16 @@ import lombok.RequiredArgsConstructor;
 @Getter
 @RequiredArgsConstructor(access = AccessLevel.PROTECTED)
 public class CountResponse {
-    private final int testerCount;
-    private final int makerCount;
-    private final int continueTestCount;
-    private final int completeTestCount;
+    private final long testerCount;
+    private final long makerCount;
+    private final long continueTestCount;
+    private final long completeTestCount;
 
     public static CountResponse newInstance(
-        final int testerCount,
-        final int makerCount,
-        final int continueTestCount,
-        final int completeTestCount
+        final long testerCount,
+        final long makerCount,
+        final long continueTestCount,
+        final long completeTestCount
     ) {
         CountResponse response = new CountResponse(
                 testerCount,
